Cleanup unused agave endpoints.
