+++ server/routes.js
@@ -2,7 +2,6 @@
 
 import express from 'express'
 import users from './handlers/users'
-import jobs from './handlers/jobs'
 import awsJobs from './handlers/awsJobs'
 import eventLogs from './handlers/eventLogs'
 import validation from './handlers/validation'
@@ -130,18 +129,13 @@ const routes = [
     method: 'get',
     url: '/datasets/:datasetId/jobs',
     middleware: [auth.datasetAccess({ optional: true })],
-    handler: jobs.getDatasetJobs,
+    handler: awsJobs.getDatasetJobs,
   },
   {
     method: 'delete',
     url: '/datasets/:datasetId/jobs',
     middleware: [auth.datasetAccess()],
-    handler: jobs.deleteDatasetJobs,
-  },
-  {
-    method: 'post',
-    url: '/jobs/:jobId/results',
-    handler: jobs.postResults,
+    handler: awsJobs.deleteDatasetJobs,
   },
   {
     method: 'get',
@@ -175,13 +169,13 @@ const routes = [
     method: 'get',
     url: '/datasets/:datasetId/jobs/:jobId/results/ticket',
     middleware: [auth.datasetAccess()],
-    handler: jobs.getDownloadTicket,
+    handler: awsJobs.getDownloadTicket,
   },
   {
     method: 'get',
     url: '/jobs',
     middleware: [auth.optional],
-    handler: jobs.getJobs,
+    handler: awsJobs.getJobs,
   },
   {
     method: 'get',
